Update loaders.py
cpu/xpu统一以fp32精度读取模型权重就可以运行了, 避免模型权重读取问题

+++ audiocraft/models/loaders.py
@@ -107,7 +107,9 @@ def load_lm_model(file_or_url_or_id: tp.Union[Path, str], device='cpu', cache_di
     if cfg.device == 'cpu':
         cfg.dtype = 'float32'
     else:
-        cfg.dtype = 'float16'
+        #cfg.dtype = 'float16'
+        #cpu/xpu统一以fp32精度读取模型权重
+        cfg.dtype = 'float32'
     _delete_param(cfg, 'conditioners.self_wav.chroma_stem.cache_path')
     _delete_param(cfg, 'conditioners.args.merge_text_conditions_p')
     _delete_param(cfg, 'conditioners.args.drop_desc_p')
